Files Added Manually
Files added are all of the files that I made changes to. I tried working on both the Video Volume issue and the Word Spacing issue, but I have become stuck in the coding. I want to supply the code for everyone so they may look at it and maybe offer some solutions.